[spec-lang] Replaced deref inference with value op and added unsigned ops.
1. Replaced derefence inference with a value operator. The value
operator will infer the type and dereference the pointer location.

The reason why I decided to go with this is because unlike the compiler
which refers to the exact pointer location or value location of a
variable, we cannot point to a value location.

E.g. If we have a variable `a` typed `struct A`, we currently do not
refer to the value of `a` because this would require the the entire
struct to be dereferenced (and the size of a struct can be large).
The current memory model is inefficient in this sense.

I am currently investigating a few ways to improve the memory models
to make this work.

2. Added unsigned comparison operators.