[constant propgation] Implemented constant propagtion.
This is the first step for data section memory abstraction.
The next step is to replace each memory access with a variable that
represents the value being accessed.
Hopefully the stack is left in the mem variable. If possible, we may
only need a stack variable for each function call (if stack is balanced
accrossed function calls).